rename ignore_field.rs -> attached_token.rs
