+++ src/parser/mod.rs
@@ -24,7 +24,7 @@ use core::{
     fmt::{self, Display},
     str::FromStr,
 };
-use helpers::ignore_field::AttachedToken;
+use helpers::attached_token::AttachedToken;
 
 use log::debug;
